Program Locator Map Prototype
These files are a prototype of the map view, including a sidebar
template for filter controls to be added in later.

The general categories currently on the sidebar represent a possible
pop-up or slide-out box with the specific program category checkboxes
once clicked.

Also added a Transit Directions button to each program inbox on the
map. This will then open u a window that will allow the user to get
transit directions from their location.